WIP: Doing refactoring...Continuing in clean branch
