+++ main.go
@@ -32,45 +32,6 @@ exp ::= 0 | 1 | -1 | ...     -- Integers
      | vars                  -- Variables
 */
 
-// Values
-
-type Kind int
-
-const (
-	ValueInt  Kind = 0
-	ValueBool Kind = 1
-	Undefined Kind = 2
-)
-
-type Val struct {
-	flag Kind
-	valI int
-	valB bool
-}
-
-func mkInt(x int) Val {
-	return Val{flag: ValueInt, valI: x}
-}
-func mkBool(x bool) Val {
-	return Val{flag: ValueBool, valB: x}
-}
-func mkUndefined() Val {
-	return Val{flag: Undefined}
-}
-
-func showVal(v Val) string {
-	var s string
-	switch {
-	case v.flag == ValueInt:
-		s = Num(v.valI).pretty()
-	case v.flag == ValueBool:
-		s = Bool(v.valB).pretty()
-	case v.flag == Undefined:
-		s = "Undefined"
-	}
-	return s
-}
-
 // Types
 
 type Type int
@@ -616,7 +577,4 @@ func (prog Prog) checkProgType() bool {
 
 }
 
-func main() {
-
-	fmt.Printf("\n")
-}
+func main() {}
