+++ values.go
@@ -1,0 +1,42 @@
+package main
+
+// Everything values related goes here
+
+// Values
+
+type Kind int
+
+const (
+	ValueInt  Kind = 0
+	ValueBool Kind = 1
+	Undefined Kind = 2
+)
+
+type Val struct {
+	flag Kind
+	valI int
+	valB bool
+}
+
+func mkInt(x int) Val {
+	return Val{flag: ValueInt, valI: x}
+}
+func mkBool(x bool) Val {
+	return Val{flag: ValueBool, valB: x}
+}
+func mkUndefined() Val {
+	return Val{flag: Undefined}
+}
+
+func showVal(v Val) string {
+	var s string
+	switch {
+	case v.flag == ValueInt:
+		s = Num(v.valI).pretty()
+	case v.flag == ValueBool:
+		s = Bool(v.valB).pretty()
+	case v.flag == Undefined:
+		s = "Undefined"
+	}
+	return s
+}
